backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
Pillow>=10.0
# HEIC/HEIF support — iPhone photos upload as HEIC by default
pillow-heif>=0.16.0
numpy>=1.26.0
scikit-learn>=1.3.2
joblib>=1.3.2
werkzeug>=3.0.1
PyJWT>=2.8.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Test dependencies (integration tests use requests against Docker backend)
pytest>=7.0.0
pytest-flask>=1.3.0
requests>=2.28.0

# Deep Learning / AI
torch>=2.0.0
torchvision>=0.15.0
# Pin LightGlue for reproducible installs
lightglue @ https://github.com/cvg/LightGlue/archive/refs/tags/v0.2.zip
einops
kornia

# Computer Vision (Use headless for Docker/Linux compatibility)
opencv-python-headless>=4.8.0